+++ custom_components/solarcharger/config/config_options_flow.py
@@ -49,6 +49,7 @@ from ..const import (
     NUMBER_CHARGER_MAX_SPEED,
     NUMBER_CHARGER_MIN_CURRENT,
     NUMBER_CHARGER_MIN_WORKABLE_CURRENT,
+    NUMBER_CHARGER_MIN_WORKABLE_CURRENT_ENTER_PAUSE_PERCENT,
     NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
     NUMBER_CHARGER_POWER_ALLOCATION_WEIGHT,
     NUMBER_CHARGER_PRIORITY,
@@ -324,11 +325,16 @@ class ConfigOptionsFlowHandler(OptionsFlow):
             self._optional(
                 subentry, NUMBER_POWER_MONITOR_DURATION, use_default
             ): NUMBER_ENTITY_SELECTOR,
-            self._optional(
-                subentry,
-                NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
-                use_default,
-            ): NUMBER_ENTITY_SELECTOR,
+            # self._optional(
+            #     subentry,
+            #     NUMBER_CHARGER_MIN_WORKABLE_CURRENT_ENTER_PAUSE_PERCENT,
+            #     use_default,
+            # ): NUMBER_ENTITY_SELECTOR,
+            # self._optional(
+            #     subentry,
+            #     NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
+            #     use_default,
+            # ): NUMBER_ENTITY_SELECTOR,
         }
 
     # ----------------------------------------------------------------------------
@@ -367,6 +373,28 @@ class ConfigOptionsFlowHandler(OptionsFlow):
             ): choose_selector(
                 api_entities,
                 NUMBER_CHARGER_MIN_WORKABLE_CURRENT,
+                NUMBER_ENTITY_SELECTOR_READ_ONLY,
+                NUMBER_ENTITY_SELECTOR,
+                modifiable_if_solarcharger_entity=True,
+            ),
+            self._optional(
+                subentry,
+                NUMBER_CHARGER_MIN_WORKABLE_CURRENT_ENTER_PAUSE_PERCENT,
+                use_default,
+            ): choose_selector(
+                api_entities,
+                NUMBER_CHARGER_MIN_WORKABLE_CURRENT_ENTER_PAUSE_PERCENT,
+                NUMBER_ENTITY_SELECTOR_READ_ONLY,
+                NUMBER_ENTITY_SELECTOR,
+                modifiable_if_solarcharger_entity=True,
+            ),
+            self._optional(
+                subentry,
+                NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
+                use_default,
+            ): choose_selector(
+                api_entities,
+                NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
                 modifiable_if_solarcharger_entity=True,