+++ custom_components/solarcharger/number.py
@@ -41,6 +41,7 @@ from .const import (
     NUMBER_CHARGER_MAX_SPEED,
     NUMBER_CHARGER_MIN_CURRENT,
     NUMBER_CHARGER_MIN_WORKABLE_CURRENT,
+    NUMBER_CHARGER_MIN_WORKABLE_CURRENT_ENTER_PAUSE_PERCENT,
     NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
     NUMBER_CHARGER_POWER_ALLOCATION_WEIGHT,
     NUMBER_CHARGER_PRIORITY,
@@ -263,6 +264,30 @@ CONFIG_NUMBER_LIST: tuple[
             native_step=0.01,
         ),
     ),
+    (
+        NUMBER_CHARGER_MIN_WORKABLE_CURRENT_ENTER_PAUSE_PERCENT,
+        SolarChargerEntityType.TYPE_LOCAL,
+        NumberEntityDescription(
+            key=NUMBER_CHARGER_MIN_WORKABLE_CURRENT_ENTER_PAUSE_PERCENT,
+            entity_category=EntityCategory.CONFIG,
+            native_unit_of_measurement=PERCENTAGE,
+            native_min_value=-500,
+            native_max_value=500,
+            native_step=1,
+        ),
+    ),
+    (
+        NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
+        SolarChargerEntityType.TYPE_LOCAL,
+        NumberEntityDescription(
+            key=NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
+            entity_category=EntityCategory.CONFIG,
+            native_unit_of_measurement=PERCENTAGE,
+            native_min_value=-500,
+            native_max_value=500,
+            native_step=1,
+        ),
+    ),
     # Control entity
     (
         NUMBER_CHARGER_PRIORITY,
@@ -435,18 +460,6 @@ CONFIG_NUMBER_LIST: tuple[
             native_min_value=0.0,
             native_max_value=60.0,
             native_step=0.01,
-        ),
-    ),
-    (
-        NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
-        SolarChargerEntityType.TYPE_LOCALHIDDEN_GLOBAL,
-        NumberEntityDescription(
-            key=NUMBER_CHARGER_MIN_WORKABLE_CURRENT_EXIT_PAUSE_PERCENT,
-            entity_category=EntityCategory.CONFIG,
-            native_unit_of_measurement=PERCENTAGE,
-            native_min_value=0,
-            native_max_value=500,
-            native_step=1,
         ),
     ),
     #####################################